Use `read_unaligned` when reading `u64`
Casting an arbitrary pointer to `u64` is an undefined behavior.

Not sure if casting of the discriminator value is really necessary, as
it is just a `u8` in this example.  But using the same
`*ptr.cast::<uXX>()` pattern in there seems wrong.  It is a noop for
`u8`, but if a `u64` is used, it becomes undefined behavior as well.

Also, the `unsafe` usage is somewhat unclear.  Dereferencing a pointer
is an unsafe operation, but the first two examples do not have `unsafe`.
Yet in the last example, when a slice is constructed, `unsafe` is
present.  I assume that the first two examples are expected to be in an
unsafe context already?